requirements.txt
pydantic-settings
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
dataiku-api-client==14.1.3
dataiku-internal-client @ http://localhost:11000/public/packages/dataiku-internal-client.tar.gz
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
geographiclib==2.1
geopy==2.4.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grandalf==0.8
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.27
langchain-core==0.3.77
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-cli==0.4.4
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.32
lark==1.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.6
openai==2.1.0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.3
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.11.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xxhash==3.6.0
zstandard==0.25.0
